Updated SceneSwitcher example to work when imported via Package Manager
When importing via Package Manager by clicking on "Import" sample
button, the path to the imported sample will contain the current
version of the package, so the code that loads the scene must be able
to load the scene without knowing its path.

Scene files were renamed to avoid potential clashing with other scenes
with same name.